backend/requirements.txt
fastapi>=0.110
uvicorn>=0.29
anthropic>=0.25
google-api-python-client>=2.120
youtube-transcript-api>=0.6
python-dotenv>=1.0
pydantic>=2.0
